Refactor logging to work well with stackdriver
