+++ Dockerfile
@@ -1,5 +1,5 @@
-ARG MONORERPO_BUILD_VERSION=latest
-FROM celo-monorepo-build-base:$MONORERPO_BUILD_VERSION
+ARG MONOREPO_BUILD_VERSION=latest
+FROM celo-monorepo-build-base:$MONOREPO_BUILD_VERSION
 
 ARG KOMENCI_VERSION
 ENV KOMENCI_VERSION=$KOMENCI_VERSION
