+++ apps/onboarding/src/config/app.config.ts
@@ -12,6 +12,7 @@ export const appConfig = registerAs('app', () => ({
   relayerRpcTimeoutMs: parseInt(process.env.RELAYER_RPC_TIMEOUT_MS, 10) || 5000,
   callbackUrl: process.env.PUBLIC_URL || 'http://localhost:3000',
   version: process.env.KOMENCI_VERSION || 'version-missing',
+  service: process.env.KOMENCI_SERVICE || 'komenci-api',
 }))
 
 export type AppConfig = ConfigType<typeof appConfig>
