+++ apps/onboarding/src/dto/SubmitMetaTransactionDto.ts
@@ -1,6 +1,5 @@
 import { IsCeloAddress } from '@app/onboarding/utils/validators'
-import { RawTransaction } from '@celo/contractkit/lib/wrappers/MetaTransactionWallet'
-import { IsHexadecimal, IsNotEmpty, IsNumberString } from 'class-validator'
+import { IsHexadecimal } from 'class-validator'
 
 export class SubmitMetaTransactionDto {
   @IsCeloAddress()
