+++ apps/relayer/src/app.module.ts
@@ -7,6 +7,7 @@ import { NetworkConfig, networkConfig } from '@app/utils/config/network.config'
 import { HttpModule, Module } from '@nestjs/common'
 import { ConfigModule, ConfigService } from '@nestjs/config'
 import { BalanceService } from 'apps/relayer/src/chain/balance.service'
+import { loggerConfigFactory } from 'apps/relayer/src/logger-config.factory'
 import { OdisService } from 'apps/relayer/src/odis/odis.service'
 import { AppController } from './app.controller'
 import { TransactionService } from './chain/transaction.service'
@@ -51,53 +52,7 @@ import { metaTransactionWalletProvider } from './contracts/MetaTransactionWallet
     KomenciLoggerModule.forRootAsync({
       providers: [ConfigService],
       inject: [ConfigService],
-      useFactory: (config: ConfigService) => {
-        const appCfg = config.get<AppConfig>('app')
-        const walletCfg = config.get<WalletConfig>('wallet')
-
-        const levelToSeverity = {
-          trace: 'DEBUG',
-          debug: 'DEBUG',
-          info: 'INFO',
-          warn: 'WARNING',
-          error: 'ERROR',
-          fatal: 'CRITICAL',
-        }
-
-        return {
-          formatters: {
-            level(label: string) {
-              const pinoLevel = label
-              const severity = levelToSeverity[pinoLevel]
-              // `@type` property tells Error Reporting to track even if there is no `stack_trace`
-              // you might want to make this an option the plugin, in our case we do want error reporting for all errors, with or without a stack
-              const typeProp =
-                pinoLevel === 'error' || pinoLevel === 'fatal'
-                  ? {
-                    '@type':
-                      'type.googleapis.com/google.devtools.clouderrorreporting.v1beta1.ReportedErrorEvent',
-                  }
-                  : {}
-              return { severity, ...typeProp }
-
-            },
-          },
-          base: {
-            ['logging.googleapis.com/labels']: {
-              service: 'relayer',
-              version: appCfg.version
-            }
-          },
-          pinoHttp: {
-            mixin: () => ({
-              relayer: walletCfg.address
-            }),
-            messageKey: 'message',
-            level: appCfg.logLevel,
-            prettyPrint: process.env.NODE_ENV !== 'production'
-          }
-        }
-      }
+      useFactory: loggerConfigFactory
     }),
     HttpModule
   ],
