+++ apps/relayer/src/dto/RawTransactionDto.ts
@@ -1,6 +1,6 @@
 import { IsCeloAddress } from '@app/onboarding/utils/validators'
 import { RawTransaction } from '@celo/contractkit/lib/wrappers/MetaTransactionWallet'
-import { IsHexadecimal, IsNotEmpty, IsNumberString } from 'class-validator'
+import { IsHexadecimal, IsNumberString } from 'class-validator'
 
 export class RawTransactionDto implements RawTransaction {
   @IsCeloAddress()
